tests/wpt/web-platform-tests/trusted-types/DOMParser.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  test(t => {
    var html = TrustedHTML.escape(STRINGS.unescapedHTML);

    var parser = new DOMParser();
    var doc = parser.parseFromString(html, "text/html");
    assert_equals(doc.body.innerText, STRINGS.unescapedHTML);
  }, "innerTextOfDoc = TrustedHTML.escape().");

  test(t => {
    var html = TrustedHTML.unsafelyCreate(STRINGS.unescapedHTML);

    var parser = new DOMParser();
    var doc = parser.parseFromString(html, "text/html");
    assert_equals(doc.body.innerText, STRINGS.unescapedText);
  }, "innerTextOfDoc = TrustedHTML.unsafelyCreate().");

  test(t => {
    var parser = new DOMParser();
    var doc = parser.parseFromString(null, "text/html");
    assert_equals(doc.body.innerText, "null");
  }, "innerTextOfDoc = null.");
</script>
